Last arranges to router and tests #2 #6 #11
