+++ src/helper/functions.ts
@@ -1,5 +1,7 @@
+import { userModel } from '../user/user.model';
+import { IUser } from '../user/user.interface';
 
-export function createUsers(numUsers: number) {
+export function createJsonUsers(numUsers: number) {
   const testUsers = [];
   for (let i = 0; i < numUsers; i++) {
     const user = {
@@ -10,6 +12,22 @@ export function createUsers(numUsers: number) {
       name: 'User' + i,
       rootFolder: '/Path/To/Root/Folder' + i,
     };
+    testUsers.push(user);
+  }
+  return testUsers;
+}
+
+export function createUsers(numUsers: number) {
+  const testUsers = [];
+  for (let i = 0; i < numUsers; i++) {
+    const user = new userModel({
+      _id: 10 * numUsers + i,
+      uniqueID: 'uID' + i,
+      creationDate: new Date(),
+      heirarchy: 'Aman/Sapir/MadorHaim/' + i,
+      name: 'User' + i,
+      rootFolder: '/Path/To/Root/Folder' + i,
+    });
     testUsers.push(user);
   }
   return testUsers;
